sämtliche SDL-Key-Warnings behoben, Typecasts eingefügt, Render(float, float) mit einer (int, int)-variante im Monster überladen. Restliche Warnungen alle wegen nicht vererbten Monstern (haben noch keine int-Render), werde ich demnächst beheben

+++ Remarius/source/Text.cpp
@@ -74,9 +74,9 @@ void CText::SetText (const char *sentence)									 // eigentlicher Text wird üb
 
 void CText::Render (float TextxPos, float TextyPos)																	
 {
-	m_TextRect.x = static_cast<float>(TextxPos);
+	m_TextRect.x = static_cast<int>(TextxPos);
 	//m_TextRect.y = static_cast<Sint16>(TextyPos);
-	m_TextRect.y = static_cast<float>(TextyPos);
+	m_TextRect.y = static_cast<int>(TextyPos);
 	m_TextRect.w = text_width;
 	m_TextRect.h = text_height;
